feat(arc-1090): move docker for tests to tests folder, own package.json
