+++ .gitignore
@@ -46,3 +46,5 @@ tests/test_results/junit.xml
 # End to end testing folders
 tests/test-results/
 tests/videos/
+tests/node_modules
+*.local.yml
