+++ openshift/Jenkinsfile
@@ -175,7 +175,7 @@ pipeline {
        loadEnvironmentVariablesFromFile('.env')
      }
       sh 'echo ${TEST_ACM_DASHBOARD_USERNAME}'
-      docker.image('mcr.microsoft.com/playwright:v1.25.0').inside('-e TEST_CLIENT_ENDPOINT=${TEST_CLIENT_ENDPOINT} -e TEST_PROXY_ENDPOINT=${TEST_PROXY_ENDPOINT}  -e TEST_OBJECT_DETAIL_PAGE_VRT=${TEST_OBJECT_DETAIL_PAGE_VRT} -e TEST_OBJECT_DETAIL_PAGE_VLAAMS_PARLEMENT=${TEST_OBJECT_DETAIL_PAGE_VLAAMS_PARLEMENT} -e TEST_VISITOR_ACCOUNT_USERNAME:${TEST_VISITOR_ACCOUNT_USERNAME} -e TEST_VISITOR_ACCOUNT_PASSWORD=${TEST_VISITOR_ACCOUNT_PASSWORD} -e TEST_CP_ADMIN_ACCOUNT_USERNAME=${TEST_CP_ADMIN_ACCOUNT_USERNAME} -e TEST_CP_ADMIN_ACCOUNT_PASSWORD=${TEST_CP_ADMIN_ACCOUNT_PASSWORD} -e TEST_MEEMOO_ADMIN_ACCOUNT_USERNAME=${TEST_MEEMOO_ADMIN_ACCOUNT_USERNAME} -e TEST_MEEMOO_ADMIN_ACCOUNT_PASSWORD=${TEST_MEEMOO_ADMIN_ACCOUNT_PASSWORD} -e TEST_ACM_DASHBOARD=${TEST_ACM_DASHBOARD} -e TEST_ACM_DASHBOARD_USERNAME=${TEST_ACM_DASHBOARD_USERNAME} -e TEST_ACM_DASHBOARD_PASSWORD=${TEST_ACM_DASHBOARD_PASSWORD}')
+      docker.image('mcr.microsoft.com/playwright:v1.25.0').inside('-e TEST_CLIENT_ENDPOINT=${TEST_CLIENT_ENDPOINT} -e TEST_PROXY_ENDPOINT=${TEST_PROXY_ENDPOINT}  -e TEST_OBJECT_DETAIL_PAGE_VRT=${TEST_OBJECT_DETAIL_PAGE_VRT} -e TEST_OBJECT_DETAIL_PAGE_VLAAMS_PARLEMENT=${TEST_OBJECT_DETAIL_PAGE_VLAAMS_PARLEMENT} -e TEST_VISITOR_ACCOUNT_USERNAME=${TEST_VISITOR_ACCOUNT_USERNAME} -e TEST_VISITOR_ACCOUNT_PASSWORD=${TEST_VISITOR_ACCOUNT_PASSWORD} -e TEST_CP_ADMIN_ACCOUNT_USERNAME=${TEST_CP_ADMIN_ACCOUNT_USERNAME} -e TEST_CP_ADMIN_ACCOUNT_PASSWORD=${TEST_CP_ADMIN_ACCOUNT_PASSWORD} -e TEST_MEEMOO_ADMIN_ACCOUNT_USERNAME=${TEST_MEEMOO_ADMIN_ACCOUNT_USERNAME} -e TEST_MEEMOO_ADMIN_ACCOUNT_PASSWORD=${TEST_MEEMOO_ADMIN_ACCOUNT_PASSWORD} -e TEST_ACM_DASHBOARD=${TEST_ACM_DASHBOARD} -e TEST_ACM_DASHBOARD_USERNAME=${TEST_ACM_DASHBOARD_USERNAME} -e TEST_ACM_DASHBOARD_PASSWORD=${TEST_ACM_DASHBOARD_PASSWORD}')
         {
          git branch: 'develop',
            credentialsId: 'meemoo-ci',
