+++ package.json
@@ -21,20 +21,20 @@
 		"test:e2e:install-chrome": "playwright install chrome",
 		"test:e2e": "npx playwright test",
 		"test:e2e:debug": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1",
-		"test:e2e:debug:t01": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t01-test-registratie-eerste-keer-inloggen-basisgebruiker.spec.ts",
-		"test:e2e:debug:t02": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t02-test-inloggen-cp-admin.spec.ts",
-		"test:e2e:debug:t03": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t03-test-inloggen-meemoo-admin-toegang-aanvragen-tot-bezoekersruimte.spec.ts",
-		"test:e2e:debug:t04": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t04-test-inloggen-bestaande-basisgebruiker.spec.ts",
-		"test:e2e:debug:t05": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t05-test-zoeken-naar-toegang-aanvragen-tot-een-bezoekersruimte.spec.ts",
-		"test:e2e:debug:t06": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t06-test-feedbackbutton-niet-ingelogd.spec.ts",
-		"test:e2e:debug:t07": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t07-test-faq-raadplegen-niet-ingelogd.spec.ts",
-		"test:e2e:debug:t08": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t08-test-toegangsaanvraag-accepteren-weigeren-door-cp-admin.spec.ts",
-		"test:e2e:debug:t09": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t09-test-mappen-profielpagina-basisgebruiker.spec.ts",
-		"test:e2e:debug:t10": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t10-test-actieve-toegang-basisgebruiker.spec.ts",
-		"test:e2e:debug:t11": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t11-test-detailpagina-object.spec.ts",
-		"test:e2e:debug:t12": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t12-test-einde-sessie-bezoekertool.spec.ts",
-		"test:e2e:debug:t13": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t13-test-instellingen-cp-beheer.spec.ts",
-		"test:e2e:debug:t14": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t14-meemoo-beheer-basis-beheersfunctionaliteiten.spec.ts",
+		"test:e2e:debug:t01": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/tests/t01-test-registratie-eerste-keer-inloggen-basisgebruiker.spec.ts",
+		"test:e2e:debug:t02": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/tests/t02-test-inloggen-cp-admin.spec.ts",
+		"test:e2e:debug:t03": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/tests/t03-test-inloggen-meemoo-admin-toegang-aanvragen-tot-bezoekersruimte.spec.ts",
+		"test:e2e:debug:t04": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/tests/t04-test-inloggen-bestaande-basisgebruiker.spec.ts",
+		"test:e2e:debug:t05": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/tests/t05-test-zoeken-naar-toegang-aanvragen-tot-een-bezoekersruimte.spec.ts",
+		"test:e2e:debug:t06": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/tests/t06-test-feedbackbutton-niet-ingelogd.spec.ts",
+		"test:e2e:debug:t07": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/tests/t07-test-faq-raadplegen-niet-ingelogd.spec.ts",
+		"test:e2e:debug:t08": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/tests/t08-test-toegangsaanvraag-accepteren-weigeren-door-cp-admin.spec.ts",
+		"test:e2e:debug:t09": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/tests/t09-test-mappen-profielpagina-basisgebruiker.spec.ts",
+		"test:e2e:debug:t10": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/tests/t10-test-actieve-toegang-basisgebruiker.spec.ts",
+		"test:e2e:debug:t11": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/tests/t11-test-detailpagina-object.spec.ts",
+		"test:e2e:debug:t12": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/tests/t12-test-einde-sessie-bezoekertool.spec.ts",
+		"test:e2e:debug:t13": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/tests/t13-test-instellingen-cp-beheer.spec.ts",
+		"test:e2e:debug:t14": "cross-env PWDEBUG=console dotenv -e .env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/tests/t14-meemoo-beheer-basis-beheersfunctionaliteiten.spec.ts",
 		"type-check": "tsc --noEmit",
 		"prepare": "husky install",
 		"storybook": "start-storybook -p 6006",
