+++ tests/package.json
@@ -1,0 +1,39 @@
+{
+	"name": "hetarchief-client-end-2-end-tests",
+	"version": "1.1.0",
+	"private": true,
+	"contributors": [
+		"Bert Verhelst <bert.verhelst@studiohyperdrive.be>"
+	],
+	"scripts": {
+		"test:e2e:install-chrome": "playwright install chrome",
+		"test:e2e": "npx playwright test",
+		"test:e2e:debug": "cross-env PWDEBUG=console dotenv -e ../.env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1",
+		"test:e2e:debug:t01": "cross-env PWDEBUG=console dotenv -e ../.env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t01-test-registratie-eerste-keer-inloggen-basisgebruiker.spec.ts",
+		"test:e2e:debug:t02": "cross-env PWDEBUG=console dotenv -e ../.env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t02-test-inloggen-cp-admin.spec.ts",
+		"test:e2e:debug:t03": "cross-env PWDEBUG=console dotenv -e ../.env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t03-test-inloggen-meemoo-admin-toegang-aanvragen-tot-bezoekersruimte.spec.ts",
+		"test:e2e:debug:t04": "cross-env PWDEBUG=console dotenv -e ../.env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t04-test-inloggen-bestaande-basisgebruiker.spec.ts",
+		"test:e2e:debug:t05": "cross-env PWDEBUG=console dotenv -e ../.env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t05-test-zoeken-naar-toegang-aanvragen-tot-een-bezoekersruimte.spec.ts",
+		"test:e2e:debug:t06": "cross-env PWDEBUG=console dotenv -e ../.env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t06-test-feedbackbutton-niet-ingelogd.spec.ts",
+		"test:e2e:debug:t07": "cross-env PWDEBUG=console dotenv -e ../.env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t07-test-faq-raadplegen-niet-ingelogd.spec.ts",
+		"test:e2e:debug:t08": "cross-env PWDEBUG=console dotenv -e ../.env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t08-test-toegangsaanvraag-accepteren-weigeren-door-cp-admin.spec.ts",
+		"test:e2e:debug:t09": "cross-env PWDEBUG=console dotenv -e ../.env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t09-test-mappen-profielpagina-basisgebruiker.spec.ts",
+		"test:e2e:debug:t10": "cross-env PWDEBUG=console dotenv -e ../.env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t10-test-actieve-toegang-basisgebruiker.spec.ts",
+		"test:e2e:debug:t11": "cross-env PWDEBUG=console dotenv -e ../.env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t11-test-detailpagina-object.spec.ts",
+		"test:e2e:debug:t12": "cross-env PWDEBUG=console dotenv -e ../.env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t12-test-einde-sessie-bezoekertool.spec.ts",
+		"test:e2e:debug:t13": "cross-env PWDEBUG=console dotenv -e ../.env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t13-test-instellingen-cp-beheer.spec.ts",
+		"test:e2e:debug:t14": "cross-env PWDEBUG=console dotenv -e ../.env.local npx playwright test --browser=chrome --headed --workers=1 --retries=0 --maxFailures=1 tests/t14-meemoo-beheer-basis-beheersfunctionaliteiten.spec.ts"
+	},
+	"dependencies": {
+	},
+	"devDependencies": {
+    "dotenv-cli": "5.0.0",
+    "cross-env": "7.0.3",
+    "uuidv4": "6.2.13",
+    "lodash": "^4.17.21",
+    "date-fns": "2.28.0",
+		"@playwright/test": "^1.22.2",
+    "ts-node": "10.4.0",
+		"typescript": "4.5.2"
+	}
+}
