+++ tests/tests/t02-test-inloggen-cp-admin.spec.ts
@@ -1,8 +1,8 @@
 import { expect, test } from '@playwright/test';
 
-import { acceptCookies } from './helpers/accept-cookies';
-import { acceptTos } from './helpers/accept-tos';
-import { loginUserMeemooIdp } from './helpers/login-user-meemoo-idp';
+import { acceptCookies } from '../helpers/accept-cookies';
+import { acceptTos } from '../helpers/accept-tos';
+import { loginUserMeemooIdp } from '../helpers/login-user-meemoo-idp';
 
 test('T02: Test inloggen CP-admin', async ({ page, context }) => {
 	// Go to the hetarchief homepage
