+++ tests/tests/t12-test-einde-sessie-bezoekertool.spec.ts
@@ -2,12 +2,12 @@ import { expect, request, test } from '@playwright/test';
 import addMinutes from 'date-fns/addMinutes';
 import { kebabCase, trim } from 'lodash';
 
-import { acceptCookies } from './helpers/accept-cookies';
-import { checkToastMessage } from './helpers/check-toast-message';
-import { fillRequestVisitBlade } from './helpers/fill-request-visit-blade';
-import { logout } from './helpers/log-out';
-import { loginUserHetArchiefIdp } from './helpers/login-user-het-archief-idp';
-import { waitForSearchResults } from './helpers/wait-for-search-results';
+import { acceptCookies } from '../helpers/accept-cookies';
+import { checkToastMessage } from '../helpers/check-toast-message';
+import { fillRequestVisitBlade } from '../helpers/fill-request-visit-blade';
+import { logout } from '../helpers/log-out';
+import { loginUserHetArchiefIdp } from '../helpers/login-user-het-archief-idp';
+import { waitForSearchResults } from '../helpers/wait-for-search-results';
 
 test('T10: Test actieve toegang basisgebruiker', async ({ page, context }) => {
 	// GO to the hetarchief homepage
