+++ tests/tests/t13-test-instellingen-cp-beheer.spec.ts
@@ -1,11 +1,11 @@
 import { expect, test } from '@playwright/test';
 
-import { acceptCookies } from './helpers/accept-cookies';
-import { checkActiveSidebarNavigationItem } from './helpers/check-active-sidebar-navigation-item';
-import { checkToastMessage } from './helpers/check-toast-message';
-import { fillRequestVisitBlade } from './helpers/fill-request-visit-blade';
-import { loginUserMeemooIdp } from './helpers/login-user-meemoo-idp';
-import { selectText } from './helpers/select-text';
+import { acceptCookies } from '../helpers/accept-cookies';
+import { checkActiveSidebarNavigationItem } from '../helpers/check-active-sidebar-navigation-item';
+import { checkToastMessage } from '../helpers/check-toast-message';
+import { fillRequestVisitBlade } from '../helpers/fill-request-visit-blade';
+import { loginUserMeemooIdp } from '../helpers/login-user-meemoo-idp';
+import { selectText } from '../helpers/select-text';
 
 test('T13: Test instellingen CP-beheer', async ({ page, context }) => {
 	// Go to the hetarchief homepage
